Fix #28 Could not find maven executable
It seems maven eclipse plugin execute maven by providing a maven.home
property resolving to a directory that does not contain bin/mvn or
bin/mvnDebug executable. This prevents the pre-commit hook installation.

To manage this usecase, when we can't run the executable, fallback to
simple 'mvn' or 'mvnDebug' command. If the simple command is not executable,
fail the entire process.